requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.80.0
h11==0.13.0
httptools==0.4.0
idna==3.3
pip==21.3.1
pydantic==1.9.2
python-dotenv==0.20.0
PyYAML==6.0
setuptools==60.2.0
sniffio==1.2.0
starlette==0.19.1
tenacity==8.0.1
typing-extensions==4.3.0
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wheel==0.37.1

aiohttp~=3.8.1
attrs~=22.1.0
pytest-asyncio~=0.19.0
